+++ packages/sheets-hyper-link-ui/src/commands/operations/popup.operations.ts
@@ -17,9 +17,9 @@
 import { CommandType, DOCS_ZEN_EDITOR_UNIT_ID_KEY, type ICommand, ICommandService, IUniverInstanceService } from '@univerjs/core';
 import { getSheetCommandTarget, type ISheetCommandSharedParams, SheetsSelectionsService } from '@univerjs/sheets';
 import { IEditorBridgeService } from '@univerjs/sheets-ui';
-import { getShouldDisableCellLink } from '../../controllers/menu';
 import { SheetsHyperLinkPopupService } from '../../services/popup.service';
 import { HyperLinkEditSourceType } from '../../types/enums/edit-source';
+import { getShouldDisableCurrentCellLink } from '../../utils';
 
 export interface IOpenHyperLinkEditPanelOperationParams extends ISheetCommandSharedParams {
     row: number;
@@ -94,7 +94,7 @@ export const InsertHyperLinkToolbarOperation: ICommand = {
     type: CommandType.OPERATION,
     id: 'sheet.operation.insert-hyper-link-toolbar',
     handler(accessor) {
-        if (getShouldDisableCellLink(accessor)) {
+        if (getShouldDisableCurrentCellLink(accessor)) {
             return false;
         }
         const commandService = accessor.get(ICommandService);
